🐛 Move helper for snippets method
Ref: https://github.com/scientist-softserv/adventist-dl/issues/469
Method `render_ocr_snippets` was causing AbstractController::DoubleRenderError
() during random situations, due to including the IiifPrintHelper in
ApplicationController.

This pull request moves it into a helper module which is injected by the
install generator, which matches how it was done in LV & NNP. See
https://github.com/scientist-softserv/louisville-hyku/blob/b672d2a3213aaee84c082eb8083d8ad70733036b/app/helpers/newspaper_works_helper.rb#L3
